fix: resolve compilation errors in lyrics components
- Added missing dp import in LyricsBackground.kt
- Made extractPaletteColors function private in file and added it to FullScreenLyricsActivity.kt
- Fixed context reference in FullScreenLyricsScreen composable
- Replaced togglePlayPause() with play()/pause() calls
- Fixed shadowElevation type mismatch (Double to Float conversion)

+++ app/src/main/java/com/music/revive/presentation/screen/player/PlayerScreen.kt
@@ -328,7 +328,7 @@ fun PlayerScreen(
                                 .graphicsLayer {
                                     scaleX = albumScale * playingScale
                                     scaleY = albumScale * playingScale
-                                    shadowElevation = (albumShadow.toPx() * 1.5)
+                                    shadowElevation = (albumShadow.toPx() * 1.5).toFloat()
                                     clip = true
                                     shape = albumShape
                                     alpha = albumAlpha
